Build registration actions
